requirements.txt
django==5.1.5
redis==5.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
pillow==11.1.0
django-filter==24.3